requirements.txt
alembic
aniso8601
atomicwrites
attrs
blinker
click
colorama
decorator
Flask
Flask-Cors
Flask-JWT-Extended
Flask-Migrate
flask-restx
Flask-SQLAlchemy
greenlet
gunicorn
iniconfig
itsdangerous
Jinja2
jsonschema
Mako
MarkupSafe
packaging
pluggy
py
PyJWT
pyparsing
pyrsistent
pytest
python-dotenv
pytz
six
SQLAlchemy
tomli
validators
Werkzeug
